Linux app tor startup
