+++ public/electron.js
@@ -27,6 +27,7 @@ function getPlatform(){
 
 const isDev = (process.env.NODE_ENV == 'development');
 
+/*
 var execPath;
 if(isDev) {
     execPath = joinPath(dirname(rootPath), 'bin');
@@ -37,6 +38,38 @@ if(isDev) {
 	    execPath = joinPath(rootPath, '../bin');
     }
 }
+*/
+
+
+let resourcesPath = undefined;
+if(getPlatform() == 'linux') {
+    resourcesPath = joinPath(dirname(rootPath), 'mercury-wallet/resources');
+} else {
+   resourcesPath = joinPath(dirname(rootPath), 'resources');
+}
+let execPath = undefined;
+let torrc = undefined;
+if(process.env.NODE_ENV == 'development') {
+    console.log("dev mode");
+    execPath = joinPath(resourcesPath, getPlatform());
+    torrc = joinPath(resourcesPath, 'etc', 'torrc');
+} else {
+    if(getPlatform() == 'linux') {
+        console.log("linux prod");
+        execPath = joinPath(rootPath, '../../Resources/bin');
+    } else {
+        console.log("mac prod");
+        execPath = joinPath(rootPath, '../../../bin');
+    }
+    torrc = joinPath(execPath, '../etc/torrc');
+}
+
+console.log("torrc: " + torrc);
+console.log("exec: " + execPath);
+
+const tor_cmd = (getPlatform() === 'win') ? `${joinPath(execPath, 'Tor', 'tor')}`: `${joinPath(execPath, 'tor')}`;
+
+console.log("tor_cmd: " + tor_cmd);
 
 let mainWindow;
